% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_trelliscope.R
\name{cluster_trelliscope}
\alias{cluster_trelliscope}
\title{cluster_trelliscope}
\usage{
cluster_trelliscope(data, trans = "identity", name = "Cluster Results",
  group = "common", path = "~/trelliscope", selfContained = F)
}
\arguments{
\item{data}{Must include, at minimum, following columns:
1. AccountNumber -- Unique numeric identifier --
2. Many date columns, with weekly counts within cells -- Weekly consecutive --
3. features
4. prediction}

\item{trans}{scale of the y-axis, see ggplot2::scale_x_continuous() -- defaults to "identity" --}

\item{name}{name of the trelliscope view -- defaults to "Cluster Results" --}

\item{group}{group of trelliscope views -- defaults to "common" --}

\item{path}{the base directory of trelliscope app -- defaults to "~/trelliscope" --}

\item{selfContained}{create the individual pre-rendered panels -- defaults to FALSE --
Intended to take data in weekly long format for trelliscope use}
}
\value{
trelliscopejs
}
\description{
cluster_trelliscope
}
\examples{
test
}
