% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_point.R
\name{cut_point}
\alias{cut_point}
\title{cut_point}
\usage{
cut_point(data, movAvg = 21, startQuant = 0.2, movAvg_zeroDays = 1,
  dialBack = 0)
}
\arguments{
\item{data}{Tibble/Data Frame with the following columns:
\itemize{
\item Account Number (unique identifier)
\item Date
\item Count
}}

\item{movAvg}{Moving average
\itemize{
\item Default is \code{21}
}}

\item{startQuant}{Quantile percentage to determine start date cut_point
\itemize{
\item Default is \code{0.2}
}}

\item{movAvg_zeroDays}{Number of consecetive moving averages evaluated at zero to begin end date cut_point process
\itemize{
\item Defaults to \code{1}
}}

\item{dialBack}{Number of days to 'dialBack' to determine a new end date cutPoint
\itemize{
\item Defaults to \code{0}
}}
}
\value{
filtered tibble with \code{cut_point} algorithim performed
}
\description{
\strong{Designed to implement \code{cut_point} algorithin on column}
}
\examples{
test
}
