% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_point_trelliscope.R
\name{cut_point_trelliscope}
\alias{cut_point_trelliscope}
\title{cut_point_trelliscope}
\usage{
cut_point_trelliscope(data, movAvg = 21, trans = "identity",
  name = "cut_point Results", group = "common", path = "~/trelliscope",
  selfContained = F)
}
\arguments{
\item{data}{Tibble/Data Frame with the following columns:
\itemize{
\item Account Number (unique identifier)
\item Date
\item Count
}}

\item{movAvg}{Moving average
\itemize{
\item Default is \code{21}
}}

\item{trans}{Scale of the y-axis, see \code{ggplot2::scale_y_continuous()}
\itemize{
\item Default is \code{identity}
}}

\item{name}{Name of the trelliscope view
\itemize{
\item Default is \code{cut_point Results}
}}

\item{group}{Group of trelliscope views
\itemize{
\item Default is \code{common}
}}

\item{path}{Base directory of trelliscope app
\itemize{
\item Default is \code{~/trelliscope}
}}

\item{selfContained}{Create the individual pre-rendered panels
\itemize{
\item Default is \code{FALSE}
}}
}
\value{
trelliscopejs object
}
\description{
\strong{Intended to take long data and output trelliscope with \code{cut_point} vertical lines}
}
\examples{
test
}
