% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutPoint.R
\name{cut_Points}
\alias{cut_Points}
\title{cut_Points}
\usage{
cut_Points(data)
}
\arguments{
\item{data}{Must include, at minimun, following columns:
1. AccountNumber -- Unique numeric identifier --
2. Date -- Daily consecutive --
3. Count -- Must include all real numbers --}
}
\value{
filtered tibble
}
\description{
cut_Points
}
\examples{
test
}
