% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutPointsTrelliscope.R
\name{cutPoint_Trelliscope}
\alias{cutPoint_Trelliscope}
\title{cutPoint_trelliscope}
\usage{
cutPoint_Trelliscope(data)
}
\arguments{
\item{data}{Must include, at minimum, following columns:
1. AccountNumber -- Unique numeric identifier --
2. Date -- Daily consecutive --
3. Count -- Must include all real numbers --
Intended to take data from long SQL Server table format}
}
\value{
trelliscopejs object
}
\description{
cutPoint_trelliscope
}
\examples{
test
}
