% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{find_LadderSequence}
\alias{find_LadderSequence}
\title{find_LadderSequence}
\usage{
find_LadderSequence(data, sequence)
}
\arguments{
\item{data}{Must be vector of numbers}

\item{sequence}{Must be one of following strings:
\itemize{
\item \code{I} - Increasing
\item \code{D} - Decreasing
\item \code{IP} - Increasing Positive
\item \code{P} - Decreasing Positive
\item \code{IN} - Increasing Negative
\item \code{DN} - Decreasing Negative
}}
}
\value{
integer
}
\description{
\strong{Inteded to find largest sequence of many ladder sequences}
}
\examples{
test
}
