% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_trelliscope.R
\name{gen_trelliscope}
\alias{gen_trelliscope}
\title{gen_trelliscope}
\usage{
gen_trelliscope(data, trans = "identity", name = "Cluster Results",
  group = "common", path = "~/trelliscope", selfContained = T)
}
\arguments{
\item{data}{Must include, at minimum, following columns:
1. AccountNumber -- Unique numeric identifier --
2. Many date columns, with weekly counts within cells -- Weekly consecutive --
3. features
4. prediction}

\item{name}{name of the trelliscope view, defaults to "Cluster Results"}

\item{group}{group of trelliscope views -- defaults to "common"}

\item{path}{the base directory of trelliscope app, defaults to "~/trelliscope"}

\item{selfContained}{defaults to TRUE
Intended to take data in weekly long format for trelliscope use}

\item{log10}{scale the y-axis by log base 10, defaults to FALSE}
}
\value{
trelliscopejs
}
\description{
gen_trelliscope
}
\examples{
test
}
