% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestInterval.R
\name{nest_append_interval}
\alias{nest_append_interval}
\title{nest_append_interval}
\usage{
nest_append_interval(nestTib, rawData, type, interval)
}
\arguments{
\item{nestTib}{Nested tibble with "AccountNumber" as identifier}

\item{rawData}{Must include, at minimun, following columns:
1. AccountNumber -- Unique numeric identifier --
2. Date -- Daily consecutive --
3. Count -- Must include all real numbers --}

\item{type}{Must be one of following strings:
1. "years"
2. "months"
3. "weeks"
4. "days"}

\item{interval}{Must be positive integer excluding zero}
}
\value{
nested tibble
}
\description{
nest_append_interval
}
\examples{
test
}
