% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_generic.R
\name{nest_core}
\alias{nest_core}
\title{nest_core}
\usage{
nest_core(data, type)
}
\arguments{
\item{data}{Tibble/Data Frame with the following columns:
\itemize{
\item Account Number (unique identifier)
\item Date
\item Count
}}

\item{type}{All meaningfull specifications in
English language are supported.
Stable arguments are:
\itemize{
\item \code{day}
\item \code{week}
\item \code{month}
\item \code{year}
}}
}
\value{
nested tibble
}
\description{
\strong{Designed to append on numerous descriptive numeric cognostics as columns
then nest into a tibble with "AccountNumber" being the identifier.
Based on \code{type} parameter, will aggregate and create cognostics by:}
\itemize{
\item Day
\item Week
\item Month
\item Year
}
}
\examples{
test
}
