% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_interval.R
\name{nest_core_interval}
\alias{nest_core_interval}
\title{nest_core_interval}
\usage{
nest_core_interval(data, type, interval, divide)
}
\arguments{
\item{data}{Tibble/Data Frame with the following columns:
\itemize{
\item Account Number (unique identifier)
\item Date
\item Count
}}

\item{type}{Must be one of following strings:
\itemize{
\item \code{years}
\item \code{months}
\item \code{weeks}
\item \code{days}
}}

\item{interval}{Must be positive integer excluding zero}

\item{divide}{Must be one of following strings:
\itemize{
\item \code{A}
\item \code{R}
\item \code{L}
}}
}
\value{
nested tibble
}
\description{
\strong{Designed to create cognostics by certain interval/ratio in the past}
}
\examples{
test
}
