% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_interval.R
\name{nest_interval_unnest}
\alias{nest_interval_unnest}
\title{nest_interval_unnest}
\usage{
nest_interval_unnest(data)
}
\arguments{
\item{data}{Tibble/Data Frame with the following columns:
\itemize{
\item Account Number (unique identifier)
\item Date
\item Count
}}
}
\value{
unnested tibble
}
\description{
\strong{Designed to unnest a nested tibble, which is a list of lists}
}
\examples{
test
}
