% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_generic.R
\name{nest_todo}
\alias{nest_todo}
\title{nest_todo}
\usage{
nest_todo(data)
}
\arguments{
\item{data}{Must include, at minimum, following columns:
\enumerate{
\item AccountNumber -- Unique numeric identifier --
\item Date -- Daily consecutive --
\item Count -- Must include all real numbers --
}}
}
\value{
nested joined tibble
}
\description{
Designed to create then join four nested tibbles into one tibble
with "AccountNumber" being the identifier:
\enumerate{
\item Daily
\item Weekly
\item Monthly
\item Yearly
}
}
\examples{
test
}
