% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_weekly.R
\name{normalize_weekly}
\alias{normalize_weekly}
\title{normalize_weekly}
\usage{
normalize_weekly(data)
}
\arguments{
\item{data}{Tibble/Data Frame with the following columns:
\itemize{
\item Account Number (unique identifier)
\item Date
\item Count
}}
}
\value{
spread tibble
}
\description{
\strong{Designed to normalize at the weekly level}
}
\examples{
test
}
