% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTrelliscope.R
\name{cluster_trelliscope}
\alias{cluster_trelliscope}
\title{cluster_trelliscope}
\usage{
cluster_trelliscope(data)
}
\arguments{
\item{data}{Must include, at minimum, following columns:
1. AccountNumber -- Unique numeric identifier --
2. Date -- Daily consecutive --
3. Count -- Must include all real numbers --
Intended to take data from long SQL Server table format}
}
\value{
spread tibble
}
\description{
cluster_trelliscope
}
\examples{
test
}
