% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_trelliscope.R
\name{cluster_trelliscope}
\alias{cluster_trelliscope}
\title{cluster_trelliscope}
\usage{
cluster_trelliscope(data, trans = "identity", name = "Cluster Results",
  group = "common", path = "~/trelliscope", selfContained = FALSE)
}
\arguments{
\item{data}{Tibble/Data Frame with the following columns:
\itemize{
\item Account Number (unique identifier)
\item Date
\item Count
\item Prediction
}}

\item{trans}{Scale of the y-axis, see \code{ggplot2::scale_y_continuous()}
\itemize{
\item Default is \code{identity}
}}

\item{name}{Name of the trelliscope view
\itemize{
\item Default is \code{Cluster Results}
}}

\item{group}{Group of trelliscope views
\itemize{
\item Default is \code{common}
}}

\item{path}{Base directory of trelliscope app
\itemize{
\item Default is \code{~/trelliscope}
}}

\item{selfContained}{Create the individual pre-rendered panels
\itemize{
\item Default is \code{FALSE}
}}
}
\value{
A trelliscopejs object
}
\description{
\strong{Intented to take data in long format for trelliscopejs use}
}
\examples{
test
}
\seealso{
\itemize{
\item \code{ggplot2::scale_y_continuous()}
\item \href{https://hafen.github.io/trelliscopejs/}{trelliscopejs}
}
}
