% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_point.R
\name{cut_point}
\alias{cut_point}
\title{cut_point}
\usage{
cut_point(data, movAvg = 21, startQuant = 0.2, movAvg_zeroDays = 1,
  dialBack = 0)
}
\arguments{
\item{data}{Must include, at minimum, following columns:
\enumerate{
\item AccountNumber -- Unique numeric identifier --
\item Date -- Daily consecutive --
\item Count -- Must include all real numbers --
}}

\item{movAvg}{Moving Average -- Defaults to 21}

\item{startQuant}{Quantile percentage to determine start date cutPoint -- Defaults to 0.2}

\item{movAvg_zeroDays}{Number of consecetive moving averages evaluated at zero to begin end date cutPoint process -- Defaults to 1}

\item{dialBack}{Number of days to 'dialBack' to determine a new end date cutPoint -- Defaults to 0}
}
\value{
filtered tibble
}
\description{
cut_point
}
\examples{
test
}
