% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_point_trelliscope.R
\name{cut_point_trelliscope}
\alias{cut_point_trelliscope}
\title{cut_point_trelliscope}
\usage{
cut_point_trelliscope(data, movAvg = 21, trans = "identity",
  name = "cutPoint Results", group = "common", path = "~/trelliscope",
  selfContained = F)
}
\arguments{
\item{data}{Must include, at minimum, following columns:
\enumerate{
\item AccountNumber -- Unique numeric identifier --
\item Date -- Daily consecutive --
\item Count -- Must include all real numbers --
Intended to take data from long SQL Server table format
}}

\item{movAvg}{Moving Average -- Defaults to 21}

\item{trans}{scale of the y-axis, see ggplot2::scale_x_continuous() -- defaults to "identity" --}

\item{name}{name of the trelliscope view -- defaults to "Cluster Results" --}

\item{group}{group of trelliscope views -- defaults to "common" --}

\item{path}{the base directory of trelliscope app -- defaults to "~/trelliscope" --}

\item{selfContained}{create the individual pre-rendered panels -- defaults to FALSE --}
}
\value{
trelliscopejs object
}
\description{
cut_point_trelliscope
}
\examples{
test
}
