% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutPoint_trelliscope.R
\name{cutPoint_trelliscope}
\alias{cutPoint_trelliscope}
\title{cutPoint_trelliscope}
\usage{
cutPoint_trelliscope(data, movAvg = 21)
}
\arguments{
\item{data}{Must include, at minimum, following columns:
1. AccountNumber -- Unique numeric identifier --    
2. Date -- Daily consecutive --    
3. Count -- Must include all real numbers --    
Intended to take data from long SQL Server table format}

\item{movAvg}{Moving Average -- Defaults to 21}
}
\value{
trelliscopejs object
}
\description{
cutPoint_trelliscope
}
\examples{
test
}
