% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{find_LadderSequence}
\alias{find_LadderSequence}
\title{find_LadderSequence}
\usage{
find_LadderSequence(data, sequence)
}
\arguments{
\item{data}{Must be vector of numbers}

\item{sequence}{Must be one of following strings:
\enumerate{
\item "I" -- Increasing
\item "D" -- Decreasing
\item "IP" -- Increasing Positive
\item "DP" -- Decreasing Positive
\item "IN" -- Increasing Negative
\item "DN" -- Decreasing Negative
}}
}
\value{
integer
}
\description{
find_LadderSequence
}
\examples{
test
}
