% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{find_LadderSequence}
\alias{find_LadderSequence}
\title{find_LadderSequence}
\usage{
find_LadderSequence(data, sequence)
}
\arguments{
\item{data}{Must be vector of numbers}

\item{sequence}{Must be one of following strings:
1. "I" -- Increasing
2. "D" -- Decreasing
3. "IP" -- Increasing Positive
4. "DP" -- Decreasing Positive
5. "IN" -- Increasing Negative
6. "DN" -- Decreasing Negative}
}
\value{
integer
}
\description{
find_LadderSequence
}
\examples{
test
}
