% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{find_SignedSequence}
\alias{find_SignedSequence}
\title{find_SignedSequence}
\usage{
find_SignedSequence(data, number)
}
\arguments{
\item{data}{Must be vector of numbers}

\item{number}{Must be one of following integers:
\itemize{
\item \code{-1}
\item \code{0}
\item \code{1}
}}
}
\value{
integer
}
\description{
\strong{Inteded to find largest sequence of negative, zero, or positive integers}
}
\examples{
test
}
