% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{find_SignedSequence}
\alias{find_SignedSequence}
\title{find_SignedSequence}
\usage{
find_SignedSequence(data, number)
}
\arguments{
\item{data}{Must be vector of numbers}

\item{number}{Must be one of following integers:
1. -1 -- Negative Integers
2. 0 -- Zero Integers
3. 1 -- Positive Integers}
}
\value{
integer
}
\description{
find_SignedSequence
}
\examples{
test
}
