% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_trelliscope.R
\name{gen_trelliscope}
\alias{gen_trelliscope}
\title{gen_trelliscope}
\usage{
gen_trelliscope(data)
}
\arguments{
\item{data}{Must include, at minimum, following columns:
1. AccountNumber -- Unique numeric identifier --
2. Many date columns, with weekly counts within cells -- Weekly consecutive --
3. features
4. prediction
Intended to take data in weekly long format for trelliscope use}
}
\value{
trelliscope
}
\description{
gen_trelliscope
}
\examples{
test
}
