% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kMeans_sparkly.R
\name{kMeans_sparkly}
\alias{kMeans_sparkly}
\title{kMeans_sparkly}
\usage{
kMeans_sparkly(data)
}
\arguments{
\item{data}{Must include, at minimum, following columns:
1. AccountNumber -- Unique numeric identifier --
2. Many date columns, with weekly counts within cells -- Weekly consecutive --
Intended to take data in weekly long format}
}
\value{
spread tibble
}
\description{
kMeans_sparkly
}
\examples{
test
}
