% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_interval.R
\name{nest_append_interval}
\alias{nest_append_interval}
\title{nest_append_interval}
\usage{
nest_append_interval(nestTib, rawData, type, interval)
}
\arguments{
\item{nestTib}{Nested tibble with `AccountNumber`` as identifier}

\item{rawData}{Tibble/Data Frame with the following columns:
\itemize{
\item Account Number (unique identifier)
\item Date
\item Count
}}

\item{type}{Must be one of following strings:
\itemize{
\item \code{years}
\item \code{months}
\item \code{weeks}
\item \code{days}
}}

\item{interval}{Must be positive integer excluding zero}
}
\value{
nested tibble
}
\description{
\strong{Designed to create cognostics by certain interval/ratio in the past}
}
\examples{
test
}
