% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_interval.R
\name{nest_append_interval}
\alias{nest_append_interval}
\title{nest_append_interval}
\usage{
nest_append_interval(nestTib, rawData, type, interval)
}
\arguments{
\item{nestTib}{Nested tibble with "AccountNumber" as identifier}

\item{rawData}{Must include, at minimun, following columns:
\enumerate{
\item AccountNumber -- Unique numeric identifier --
\item Date -- Daily consecutive --
\item Count -- Must include all real numbers --
}}

\item{type}{Must be one of following strings:
\enumerate{
\item "years"
\item "months"
\item "weeks"
\item "days"
}}

\item{interval}{Must be positive integer excluding zero}
}
\value{
nested tibble
}
\description{
nest_append_interval
}
\examples{
test
}
