% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestGeneric.R
\name{nest_core}
\alias{nest_core}
\title{nest_core}
\usage{
nest_core(data, type)
}
\arguments{
\item{data}{Must include, at minimun, following columns:
1. AccountNumber -- Unique numeric identifier --
2. Date -- Daily consecutive --
3. Count -- Must include all real numbers --}

\item{type}{All meaningfull specifications in
English language are supported.
Stable arguments are:
1. day
2. week
3. month
4. year}
}
\value{
nested tibble
}
\description{
Designed to append on numerous descriptive numeric cognostics as columns
then nest into a tibble with "AccountNumber" being the identifier.
Based on `type` parameter, will aggregate and create cognostics by:
1. Day
2. Week
3. Month
4. Year
}
\examples{
test
}
