% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_generic.R
\name{nest_core}
\alias{nest_core}
\title{nest_core}
\usage{
nest_core(data, type)
}
\arguments{
\item{data}{Must include, at minimun, following columns:
\enumerate{
\item AccountNumber -- Unique numeric identifier --
\item Date -- Daily consecutive --
\item Count -- Must include all real numbers --
}}

\item{type}{All meaningfull specifications in
English language are supported.
Stable arguments are:
\enumerate{
\item day
\item week
\item month
\item year
}}
}
\value{
nested tibble
}
\description{
Designed to append on numerous descriptive numeric cognostics as columns
then nest into a tibble with "AccountNumber" being the identifier.
Based on \code{type} parameter, will aggregate and create cognostics by:
\enumerate{
\item Day
\item Week
\item Month
\item Year
}
}
\examples{
test
}
