% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_interval.R
\name{nest_core_interval}
\alias{nest_core_interval}
\title{nest_core_interval}
\usage{
nest_core_interval(data, type, interval, divide)
}
\arguments{
\item{data}{Must include, at minimun, following columns:
1. AccountNumber -- Unique numeric identifier --
2. Date -- Daily consecutive --
3. Count -- Must include all real numbers --}

\item{type}{Must be one of following strings:
1. "years"
2. "months"
3. "weeks"
4. "days"}

\item{interval}{Must be positive integer excluding zero}

\item{divide}{Must be one of following strings:
1. "A"
2. "R"
3. "L"}
}
\value{
nested tibble
}
\description{
nest_core_interval
}
\examples{
test
}
