% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_interval.R
\name{nest_core_interval}
\alias{nest_core_interval}
\title{nest_core_interval}
\usage{
nest_core_interval(data, type, interval, divide)
}
\arguments{
\item{data}{Must include, at minimun, following columns:
\enumerate{
\item AccountNumber -- Unique numeric identifier --
\item Date -- Daily consecutive --
\item Count -- Must include all real numbers --
}}

\item{type}{Must be one of following strings:
\enumerate{
\item "years"
\item "months"
\item "weeks"
\item "days"
}}

\item{interval}{Must be positive integer excluding zero}

\item{divide}{Must be one of following strings:
\enumerate{
\item "A"
\item "R"
\item "L"
}}
}
\value{
nested tibble
}
\description{
nest_core_interval
}
\examples{
test
}
