% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_interval.R
\name{nest_interval}
\alias{nest_interval}
\title{nest_interval}
\usage{
nest_interval(data, type, interval)
}
\arguments{
\item{data}{Must include, at minimum, following columns:
\enumerate{
\item AccountNumber -- Unique numeric identifier --
\item Date -- Daily consecutive --
\item Count -- Must include all real numbers --
}}

\item{type}{Must be one of following strings:
\enumerate{
\item "years"
\item "months"
\item "weeks"
\item "days"
}}

\item{interval}{Must be positive integer excluding zero}
}
\value{
nested joined tibble
}
\description{
nest_interval
}
\examples{
test
}
