% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_interval.R
\name{nest_interval_unnest}
\alias{nest_interval_unnest}
\title{nest_interval_unnest}
\usage{
nest_interval_unnest(data)
}
\arguments{
\item{data}{Must include, at minimun, following columns:
1. AccountNumber -- Unique numeric identifier --
2. Date -- Daily consecutive --
3. Count -- Must include all real numbers --}
}
\value{
unnested tibble
}
\description{
nest_interval_unnest
}
\examples{
test
}
