% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_generic.R
\name{nest_todo}
\alias{nest_todo}
\title{nest_todo}
\usage{
nest_todo(data)
}
\arguments{
\item{data}{Tibble/Data Frame with the following columns:
\itemize{
\item Account Number (unique identifier)
\item Date
\item Count
}}
}
\value{
nested and joined tibble
}
\description{
\strong{Designed to create then join four nested tibbles into one tibble
with \code{AccountNumber} being the identifier by:}
\itemize{
\item Daily
\item Weekly
\item Monthly
\item Yearly
}
}
\examples{
test
}
