% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestGeneric.R
\name{nest_todo}
\alias{nest_todo}
\title{nest_todo}
\usage{
nest_todo(data)
}
\arguments{
\item{data}{Must include, at minimum, following columns:
1. AccountNumber -- Unique numeric identifier --
2. Date -- Daily consecutive --
3. Count -- Must include all real numbers --}
}
\value{
nested joined tibble
}
\description{
Designed to create then join four nested tibbles into one tibble
with "AccountNumber" being the identifier:
1. Daily
2. Weekly
3. Monthly
4. Yearly
}
\examples{
test
}
