% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_weekly.R
\name{normalize_weekly}
\alias{normalize_weekly}
\title{normalize_weekly}
\usage{
normalize_weekly(data)
}
\arguments{
\item{data}{Must include, at minimum, following columns:
\enumerate{
\item AccountNumber -- Unique numeric identifier --
\item Date -- Daily consecutive --
\item Count -- Must include all real numbers --
Intended to take data from long SQL Server table format
}}
}
\value{
spread tibble
}
\description{
normalize_weekly
}
\examples{
test
}
